WIP on #16
It seems to work for GHDL with the mcode backend, but doesn't with the
GCC backend. It looks like GHDL can't be told to look for the previously
generated object files in a different directory, so I guess we'll have
to copy the relevant object over to the test case directory first?
Definitely need to test with the LLVM backend too, then. This is not as
trivial as I'd hoped.